app completo

+++ app/Http/Controllers/PostController.php
@@ -3,10 +3,22 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
+use DB;
 
 class PostController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except'=> ['index','show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -38,13 +50,24 @@ class PostController extends Controller
     {
         $this->validate($request,[
           'title' => 'required',
-          'body' => 'required'
+          'body' => 'required',
+          'cover_image' => 'image|nullable|max:1999'
         ]);
+        if($request->hasFile('cover_image')){
+          $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+          $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+          $extension = $request->file('cover_image')->getClientOriginalExtension();
+          $fileNameToStore = $filename.'_'.time().'.'.$extension;
+          $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        } else{
+          $fileNameToStore = 'noimage.jpg';
+        }
         //Create Post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
         return redirect('/post')->with('success', 'Post Criado');
     }
@@ -69,8 +92,14 @@ class PostController extends Controller
      */
     public function edit($id)
     {
+
           $post = Post::find($id);
-          return view('posts.edit')->with('post', $post);
+          if(auth()->user()->id !== $post->user_id){
+            return redirect('/post')->with('error', 'Não autorizado');
+          }
+              return view('posts.edit')->with('post', $post);
+
+
     }
 
     /**
@@ -86,10 +115,19 @@ class PostController extends Controller
         'title' => 'required',
         'body' => 'required'
       ]);
+      if($request->hasFile('cover_image')){
+        $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+        $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+        $extension = $request->file('cover_image')->getClientOriginalExtension();
+        $fileNameToStore = $filename.'_'.time().'.'.$extension;
+        $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+      }
       //Create Post
       $post = Post::find($id);
       $post->title = $request->input('title');
       $post->body = $request->input('body');
+      if($request->hasFile('cover_image'))
+        $post->cover_image = $fileNameToStore;
       $post->save();
       return redirect('/post')->with('success', 'Post Atualizado');
     }
@@ -103,6 +141,13 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+          return redirect('/post')->with('error', 'Não autorizado');
+        }
+
+        if($post->cover_image != 'noimage.jpg'){
+          Storage::delete('public/cover_images/'.$post->cover_image);
+        }
         $post->delete();
         return redirect('/post')->with('success', 'Post Deletado');
     }
